Reviewed-on: https://git.irmb.bau.tu-bs.de/VirtualFluids/VirtualFluids_dev/pulls/67

This pow-solution is not optimal but will work hopefully. The kernel performance will probably be decreased, because "double" calculations are 3-5 times slower compared to floating point operations.

+++ src/gpu/VirtualFluids_GPU/Kernel/Kernels/WaleKernels/Advection/Compressible/CumulantK17/WaleCumulantK17Comp_Device.cu
@@ -435,9 +435,9 @@ extern "C" __global__ void LB_Kernel_WaleCumulantK17Comp(
 			////////////////////////////////////////////////////////////////////////////////////
 			//the force be with you
 			{
-				real fx = forces[0] / (pow(c2o1, level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
-				real fy = forces[1] / (pow(c2o1, level)); //zero;
-				real fz = forces[2] / (pow(c2o1, level)); //zero;
+				real fx = forces[0] / (pow((double)c2o1, (double)level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
+				real fy = forces[1] / (pow((double)c2o1, (double)level)); //zero;
+				real fz = forces[2] / (pow((double)c2o1, (double)level)); //zero;
 				vvx += fx*c1o2;
 				vvy += fy*c1o2;
 				vvz += fz*c1o2;
